refactor: extract cloudevents functions, privateize file-global identifiers
