+++ src/cloudevents.ts
@@ -13,6 +13,9 @@
 // limitations under the License.
 
 import * as express from 'express';
+import {
+  CloudEventsContext,
+} from './functions';
 
 /**
  * Checks whether the incoming request is a CloudEvents event in binary content
@@ -31,4 +34,22 @@ export function isBinaryCloudEvent(req: express.Request): boolean {
     req.header('ce-source') &&
     req.header('ce-id')
   );
+}
+
+/**
+ * Returns a CloudEvents context from the given CloudEvents request. Context
+ * attributes are retrieved from request headers.
+ *
+ * @param req Express request object.
+ * @return CloudEvents context.
+ */
+export function getBinaryCloudEventContext(req: express.Request): CloudEventsContext {
+  const context: CloudEventsContext = {};
+  for (const name in req.headers) {
+    if (name.startsWith('ce-')) {
+      const attributeName = name.substr('ce-'.length);
+      context[attributeName] = req.header(name);
+    }
+  }
+  return context;
 }
